+++ src/character/effect/special_effect.ts
@@ -1,6 +1,6 @@
 
 namespace Charjs {
-    export class SpecialEffect extends AbstractObject {
+    export class SpecialEffect extends AbstractEffect {
         cchars = [[[0, 3, 1, 4, 0, 9], [0, 2, 1, 6, 0, 1, 1, 4, 0, 3], [0, 1, 1, 13, 0, 2], [1, 15, 0, 1], [1, 15, 0, 1], [1, 15, 0, 1], [1, 14, 0, 2], [1, 15, 0, 1], [0, 1, 1, 14, 0, 1], [0, 2, 1, 14], [0, 2, 1, 14], [0, 1, 1, 15], [0, 1, 1, 14, 0, 1], [0, 1, 1, 14, 0, 1], [0, 2, 1, 11, 0, 3], [0, 6, 1, 4, 0, 6]], [[0, 16], [0, 3, 1, 2, 0, 4, 1, 1, 0, 6], [0, 2, 1, 4, 0, 2, 1, 3, 0, 5], [0, 2, 1, 4, 0, 3, 1, 1, 0, 2, 1, 1, 0, 3], [0, 2, 1, 4, 0, 5, 1, 3, 0, 2], [0, 3, 1, 2, 0, 6, 1, 3, 0, 2], [0, 12, 1, 1, 0, 3], [0, 8, 1, 2, 0, 6], [0, 7, 1, 4, 0, 5], [0, 8, 1, 2, 0, 6], [0, 4, 1, 1, 0, 7, 1, 1, 0, 3], [0, 3, 1, 3, 0, 5, 1, 3, 0, 2], [0, 3, 1, 3, 0, 1, 1, 1, 0, 3, 1, 3, 0, 2], [0, 4, 1, 1, 0, 1, 1, 3, 0, 3, 1, 1, 0, 3], [0, 7, 1, 1, 0, 8], [0, 16]], [[0, 16], [0, 16], [0, 16], [0, 3, 1, 2, 0, 4, 1, 1, 0, 6], [0, 2, 1, 4, 0, 5, 1, 2, 0, 3], [0, 3, 1, 2, 0, 6, 1, 2, 0, 3], [0, 16], [0, 16], [0, 16], [0, 9, 1, 1, 0, 6], [0, 8, 1, 3, 0, 5], [0, 3, 1, 2, 0, 4, 1, 1, 0, 6], [0, 3, 1, 2, 0, 11], [0, 11, 1, 2, 0, 3], [0, 8, 1, 1, 0, 7], [0, 16]], [[0, 16], [0, 16], [0, 16], [0, 16], [0, 3, 1, 1, 0, 12], [0, 16], [0, 12, 1, 1, 0, 3], [0, 16], [0, 16], [0, 9, 1, 1, 0, 6], [0, 16], [0, 3, 1, 1, 0, 12], [0, 12, 1, 1, 0, 3], [0, 16], [0, 16], [0, 16]], [[0, 16], [0, 16], [0, 16], [0, 16], [0, 16], [0, 8, 2, 1, 0, 7], [0, 8, 2, 1, 0, 7], [0, 5, 2, 7, 0, 4], [0, 6, 2, 5, 0, 5], [0, 7, 2, 3, 0, 6], [0, 6, 2, 5, 0, 5], [0, 5, 2, 2, 0, 3, 2, 2, 0, 4], [0, 16], [0, 16], [0, 16], [0, 16]]];
         colors = ['', '#fff', '#fffd34'];
         chars = null;
@@ -100,6 +100,18 @@ namespace Charjs {
 
         private currentIndex = 0;
         private prevStars: HTMLCanvasElement[] = [];
+
+
+        private static staticEffect: SpecialEffect = null;
+
+        public static drawSpecial(targetDom: HTMLElement, pos: IPosition, pixSize: number) {
+            if (!SpecialEffect.staticEffect) {
+                SpecialEffect.staticEffect = new SpecialEffect(targetDom, pixSize);
+                SpecialEffect.staticEffect.init();
+            }
+            SpecialEffect.staticEffect.drawEffect(pos);
+        }
+
 
         drawEffect(pos: IPosition) {
             let tEffect = this.getTimer(() => {
