+++ src/character/object/abstract_other_object.ts
@@ -1,0 +1,5 @@
+namespace Charjs {
+    export abstract class AbstractOtherObject extends AbstractObject implements IObject {
+        entityEnemies: IEnemy[] = [];
+    }
+}
